+++ CHANGELOG.md
@@ -20,8 +20,22 @@ Whitespace conventions:
 
 
 
+
 ## [0.10.2] - Unreleased
+
+
+### Added
+
+- Support for Sprockets 3+ processor / transformer API. (No more deprecations on Sprockets 3.7+)
+
+
+### Fixed
+
 - Fixed inheritance from the `Module` class (#1476)
+
+
+
+
 
 ## [0.10.1] - 2016-07-06